requirements.txt
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.2
django-modelcluster==4.4
django-npm==1.0.0
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
Pillow==5.4.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.25.3
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
